+++ app/Bootstrap.scala
@@ -2,8 +2,8 @@
 import java.util.concurrent.Executors
 
 import actors.QueueActor
+import com.kakao.s2graph.core.utils.logger
 import com.kakao.s2graph.core.{ExceptionHandler, Graph}
-import com.kakao.s2graph.logger
 import config.Config
 import controllers.{AdminController, ApplicationController}
 import play.api.Application
