+++ app/actors/QueueActor.scala
@@ -6,7 +6,7 @@ import actors.Protocol.FlushAll
 import akka.actor._
 import com.kakao.s2graph.core.ExceptionHandler._
 import com.kakao.s2graph.core._
-import com.kakao.s2graph.logger
+import com.kakao.s2graph.core.utils.logger
 import config.Config
 import play.api.Play.current
 import play.api.libs.concurrent.Akka
