+++ app/controllers/AdminController.scala
@@ -2,7 +2,7 @@ package controllers
 
 import com.kakao.s2graph.core._
 import com.kakao.s2graph.core.mysqls._
-import com.kakao.s2graph.logger
+import com.kakao.s2graph.core.utils.logger
 import play.api.libs.json._
 import play.api.mvc
 import play.api.mvc.{Action, Controller}
