+++ app/controllers/QueryController.scala
@@ -5,7 +5,7 @@ import com.kakao.s2graph.core.GraphExceptions.BadQueryException
 import com.kakao.s2graph.core._
 import com.kakao.s2graph.core.mysqls._
 import com.kakao.s2graph.core.types.{LabelWithDirection, VertexId}
-import com.kakao.s2graph.logger
+import com.kakao.s2graph.core.utils.logger
 import config.Config
 import play.api.libs.json.{JsArray, JsObject, JsValue, Json}
 import play.api.mvc.{Action, Controller, Result}
@@ -33,12 +33,12 @@ object QueryController extends Controller with RequestParser {
 
   private def eachQuery(post: (Seq[QueryResult], Seq[QueryResult]) => JsValue)(q: Query): Future[JsValue] = {
     val filterOutQueryResultsLs = q.filterOutQuery match {
-      case Some(filterOutQuery) => Graph.getEdgesAsync(filterOutQuery)
+      case Some(filterOutQuery) => Graph.getEdges(filterOutQuery)
       case None => Future.successful(Seq.empty)
     }
 
     for {
-      queryResultsLs <- Graph.getEdgesAsync(q)
+      queryResultsLs <- Graph.getEdges(q)
       filterOutResultsLs <- filterOutQueryResultsLs
     } yield {
       val json = post(queryResultsLs, filterOutResultsLs)
@@ -86,8 +86,8 @@ object QueryController extends Controller with RequestParser {
       val q = toQuery(jsonQuery)
       val filterOutQuery = Query(q.vertices, Vector(q.steps.last))
 
-      val fetchFuture = Graph.getEdgesAsync(q)
-      val excludeFuture = Graph.getEdgesAsync(filterOutQuery)
+      val fetchFuture = Graph.getEdges(q)
+      val excludeFuture = Graph.getEdges(filterOutQuery)
 
       for {
         queryResultLs <- fetchFuture
@@ -152,8 +152,8 @@ object QueryController extends Controller with RequestParser {
       val q = toQuery(jsonQuery)
       val filterOutQuery = Query(q.vertices, Vector(q.steps.last))
 
-      val fetchFuture = Graph.getEdgesAsync(q)
-      val excludeFuture = Graph.getEdgesAsync(filterOutQuery)
+      val fetchFuture = Graph.getEdges(q)
+      val excludeFuture = Graph.getEdges(filterOutQuery)
 
       for {
         queryResultLs <- fetchFuture
@@ -185,8 +185,8 @@ object QueryController extends Controller with RequestParser {
       val q = toQuery(jsonQuery)
       val filterOutQuery = Query(q.vertices, Vector(q.steps.last))
 
-      val fetchFuture = Graph.getEdgesAsync(q)
-      val excludeFuture = Graph.getEdgesAsync(filterOutQuery)
+      val fetchFuture = Graph.getEdges(q)
+      val excludeFuture = Graph.getEdges(filterOutQuery)
 
       for {
         queryResultLs <- fetchFuture
@@ -242,10 +242,10 @@ object QueryController extends Controller with RequestParser {
           //          logger.debug(s"SrcVertex: $src")
           //          logger.debug(s"TgtVertex: $tgt")
           //          logger.debug(s"direction: $dir")
-          (src, tgt, label, dir.toInt)
+          (src, tgt, QueryParam(LabelWithDirection(label.id.get, dir)))
         }
 
-      Graph.checkEdges(quads, isInnerCall = false).map { case queryResultLs =>
+      Graph.checkEdges(quads).map { case queryResultLs =>
         val edgeJsons = for {
           queryResult <- queryResultLs
           (edge, score) <- queryResult.edgeWithScoreLs
@@ -289,7 +289,7 @@ object QueryController extends Controller with RequestParser {
         }
       }
 
-      Graph.getVerticesAsync(vertices) map { vertices =>
+      Graph.getVertices(vertices) map { vertices =>
         val json = PostProcess.verticesToJson(vertices)
         jsonResponse(json, "result_size" -> calcSize(json).toString)
       }
