+++ test/controllers/AdminControllerSpec.scala
@@ -2,9 +2,9 @@ package controllers
 
 import com.kakao.s2graph.core.mysqls.Label
 import play.api.http.HeaderNames
+import play.api.test.Helpers._
 import play.api.test.{FakeApplication, FakeRequest}
 import test.controllers.SpecCommon
-import play.api.test.Helpers._
 
 import scala.concurrent.Await
