+++ test/controllers/EdgeControllerSpec.scala
@@ -8,15 +8,15 @@ import play.api.{Application => PlayApplication}
  * Created by hsleep(honeysleep@gmail.com) on 15. 9. 1..
  */
 class EdgeControllerSpec extends PlaySpecification {
-  "EdgeControllerSpec" should {
-    implicit val app = FakeApplication()
-
-    "bad request invalid json" in new WithApplication(app) {
-      val insertUrl = "http://localhost:9000/graphs/edges/insert"
-      val req = FakeRequest("POST", insertUrl).withBody("").withHeaders(HeaderNames.CONTENT_TYPE -> "application/json")
-      val result = EdgeController.inserts().apply(req).run
-
-      status(result) must_== BAD_REQUEST
-    }
-  }
+//  "EdgeControllerSpec" should {
+//    implicit val app = FakeApplication()
+//
+//    "bad request invalid json" in new WithApplication(app) {
+//      val insertUrl = "http://localhost:9000/graphs/edges/insert"
+//      val req = FakeRequest("POST", insertUrl).withBody("").withHeaders(HeaderNames.CONTENT_TYPE -> "application/json")
+//      val result = EdgeController.inserts().apply(req).run
+//
+//      status(result) must_== BAD_REQUEST
+//    }
+//  }
 }
