+++ test/controllers/StrongLabelDeleteSpec.scala
@@ -49,99 +49,99 @@ class StrongLabelDeleteSpec extends SpecCommon {
   }
 
 
-//  "strong label delete test" should {
-//    running(FakeApplication()) {
-//      // insert bulk and wait ..
-//      val jsResult = contentAsJson(EdgeController.mutateAndPublish(bulkEdges(), withWait = true))
-//      Thread.sleep(asyncFlushInterval)
-//    }
-//
-//
-//
-//
-//    "test strong consistency select" in {
-//      running(FakeApplication()) {
-//        var result = getEdges(query(0))
-//        println(result)
-//        (result \ "results").as[List[JsValue]].size must equalTo(2)
-//        result = getEdges(query(10))
-//        println(result)
-//        (result \ "results").as[List[JsValue]].size must equalTo(2)
-//        true
-//      }
-//    }
-//
-//    "test strong consistency duration. insert -> delete -> insert" in {
-//      running(FakeApplication()) {
-//        val ts0 = 1
-//        val ts1 = 2
-//        val ts2 = 3
-//
-//        val edges = Seq(
-//          Seq(5, "insert", "edge", "-10", "-20", testLabelName2).mkString("\t"),
-//          Seq(10, "delete", "edge", "-10", "-20", testLabelName2).mkString("\t"),
-//          Seq(20, "insert", "edge", "-10", "-20", testLabelName2).mkString("\t")
-//        ).mkString("\n")
-//
-//        val jsResult = contentAsJson(EdgeController.mutateAndPublish(edges, withWait = true))
-//
-//        Thread.sleep(asyncFlushInterval)
-//        val result = getEdges(query(-10))
-//
-//        println(result)
-//
-//        true
-//      }
-//    }
-//
-//    "test strong consistency deleteAll" in {
-//      running(FakeApplication()) {
-//        val deletedAt = 100
-//        var result = getEdges(query(20, direction = "in", columnName = testTgtColumnName))
-//        println(result)
-//        (result \ "results").as[List[JsValue]].size must equalTo(3)
-//
-//
-//
-//        val json = Json.arr(Json.obj("label" -> testLabelName2,
-//          "direction" -> "in", "ids" -> Json.arr("20"), "timestamp" -> deletedAt))
-//        println(json)
-//        EdgeController.deleteAllInner(json)
-//        Thread.sleep(asyncFlushInterval)
-//
-//
-//        result = getEdges(query(11, direction = "out"))
-//        println(result)
-//        (result \ "results").as[List[JsValue]].size must equalTo(0)
-//
-//        result = getEdges(query(12, direction = "out"))
-//        println(result)
-//        (result \ "results").as[List[JsValue]].size must equalTo(0)
-//
-//        result = getEdges(query(10, direction = "out"))
-//        println(result)
-//        // 10 -> out -> 20 should not be in result.
-//        (result \ "results").as[List[JsValue]].size must equalTo(1)
-//        (result \\ "to").size must equalTo(1)
-//        (result \\ "to").head.as[String] must equalTo("21")
-//
-//
-//        result = getEdges(query(20, direction = "in", columnName = testTgtColumnName))
-//        println(result)
-//        (result \ "results").as[List[JsValue]].size must equalTo(0)
-//
-//        val jsResult = contentAsJson(EdgeController.mutateAndPublish(bulkEdges(startTs = deletedAt + 1), withWait = true))
-//        Thread.sleep(asyncFlushInterval)
-//
-//        result = getEdges(query(20, direction = "in", columnName = testTgtColumnName))
-//        println(result)
-//        (result \ "results").as[List[JsValue]].size must equalTo(3)
-//
-//        true
-//
-//      }
-//    }
-//  }
+  "strong label delete test" should {
+    running(FakeApplication()) {
+      // insert bulk and wait ..
+      val jsResult = contentAsJson(EdgeController.mutateAndPublish(bulkEdges(), withWait = true))
+      Thread.sleep(asyncFlushInterval)
+    }
+
+
+
+
+    "test strong consistency select" in {
+      running(FakeApplication()) {
+        var result = getEdges(query(0))
+        println(result)
+        (result \ "results").as[List[JsValue]].size must equalTo(2)
+        result = getEdges(query(10))
+        println(result)
+        (result \ "results").as[List[JsValue]].size must equalTo(2)
+        true
+      }
+    }
+
+    "test strong consistency duration. insert -> delete -> insert" in {
+      running(FakeApplication()) {
+        val ts0 = 1
+        val ts1 = 2
+        val ts2 = 3
+
+        val edges = Seq(
+          Seq(5, "insert", "edge", "-10", "-20", testLabelName2).mkString("\t"),
+          Seq(10, "delete", "edge", "-10", "-20", testLabelName2).mkString("\t"),
+          Seq(20, "insert", "edge", "-10", "-20", testLabelName2).mkString("\t")
+        ).mkString("\n")
+
+        val jsResult = contentAsJson(EdgeController.mutateAndPublish(edges, withWait = true))
+
+        Thread.sleep(asyncFlushInterval)
+        val result = getEdges(query(-10))
+
+        println(result)
+
+        true
+      }
+    }
+
+    "test strong consistency deleteAll" in {
+      running(FakeApplication()) {
+        val deletedAt = 100
+        var result = getEdges(query(20, direction = "in", columnName = testTgtColumnName))
+        println(result)
+        (result \ "results").as[List[JsValue]].size must equalTo(3)
+
+
+
+        val json = Json.arr(Json.obj("label" -> testLabelName2,
+          "direction" -> "in", "ids" -> Json.arr("20"), "timestamp" -> deletedAt))
+        println(json)
+        EdgeController.deleteAllInner(json)
+        Thread.sleep(asyncFlushInterval)
+
+
+        result = getEdges(query(11, direction = "out"))
+        println(result)
+        (result \ "results").as[List[JsValue]].size must equalTo(0)
+
+        result = getEdges(query(12, direction = "out"))
+        println(result)
+        (result \ "results").as[List[JsValue]].size must equalTo(0)
+
+        result = getEdges(query(10, direction = "out"))
+        println(result)
+        // 10 -> out -> 20 should not be in result.
+        (result \ "results").as[List[JsValue]].size must equalTo(1)
+        (result \\ "to").size must equalTo(1)
+        (result \\ "to").head.as[String] must equalTo("21")
+
+
+        result = getEdges(query(20, direction = "in", columnName = testTgtColumnName))
+        println(result)
+        (result \ "results").as[List[JsValue]].size must equalTo(0)
+
+        val jsResult = contentAsJson(EdgeController.mutateAndPublish(bulkEdges(startTs = deletedAt + 1), withWait = true))
+        Thread.sleep(asyncFlushInterval)
+
+        result = getEdges(query(20, direction = "in", columnName = testTgtColumnName))
+        println(result)
+        (result \ "results").as[List[JsValue]].size must equalTo(3)
+
+        true
+
+      }
+    }
+  }
   
 
   "labelargeSet of contention" should {
@@ -167,7 +167,6 @@ class StrongLabelDeleteSpec extends SpecCommon {
           }
         val bulkEdge = bulkEdgeStr.mkString("\n")
         val jsResult = contentAsJson(EdgeController.mutateAndPublish(bulkEdge, withWait = true))
-        Thread.sleep(asyncFlushInterval)
       }
       Thread.sleep(asyncFlushInterval)
       val expectedDegree = lastOps.count(op => op != "delete" && op != "none")
@@ -206,6 +205,7 @@ class StrongLabelDeleteSpec extends SpecCommon {
         val deleteAllRequest = Json.arr(Json.obj("label" -> labelName, "ids" -> Json.arr(src), "timestamp" -> deletedAt))
 
         val jsResult = contentAsString(EdgeController.deleteAllInner(deleteAllRequest))
+        Thread.sleep(asyncFlushInterval)
 
         val result = getEdges(query(id = src))
         println(result)
